A3/requirement.txt
click==8.0.4
Flask==2.0.3
requests
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
boto3==1.26.78
botocore==1.20.54
pycryptodome==3.17